Fri, Oct 13, 2023, 5:14 PM +02:00

+++ problemset2/readability/readability.c
@@ -32,8 +32,8 @@ int main(void)
 
     words++;
 
-    float L = letters / words * 100;
-    float S = sent / words * 100;
+    float L = (float) letters / words * 100;
+    float S = (float) sent / words * 100;
 
     int index = 0.0588 * L - 0.296 * S - 15.8;
